cadgen: split generation.py and step_scene.py into single-concern submodules
