+++ packages/cadgen/src/cadgen/_internal/generation.py
@@ -90,1000 +90,64 @@ from cadgen._internal.step_scene import (
     adaptive_mesh_resolution_for_scene,
     step_file_hash,
 )
-
-GIT_LFS_POINTER_PREFIX = b"version https://git-lfs.github.com/spec/v1\n"
-
-
-@dataclass(frozen=True)
-class EntrySpec:
-    source_ref: str
-    cad_ref: str
-    kind: str
-    source_path: Path
-    display_name: str
-    source: str
-    step_path: Path | None = None
-    script_path: Path | None = None
-    generator_metadata: GeneratorMetadata | None = None
-    dxf_path: Path | None = None
-    step_export_path: Path | None = None
-    dxf_export_path: Path | None = None
-    mesh_tolerance: float = DEFAULT_MESH_TOLERANCE
-    mesh_angular_tolerance: float = DEFAULT_MESH_ANGULAR_TOLERANCE
-    mesh_tolerance_explicit: bool = False
-    mesh_angular_tolerance_explicit: bool = False
-    color: tuple[float, float, float, float] | None = None
-
-    @property
-    def entry_path(self) -> Path | None:
-        # The actual on-disk ENTRY file the render cache is keyed by: the `.step.py` generator for
-        # a generated model, or the `.step`/`.stp` itself for an imported one. So a generated
-        # `<name>.step.py` and an imported `<name>.step` get distinct cache packages.
-        return self.script_path if self.script_path is not None else self.step_path
-
-
-@dataclass
-class GeneratedStepResult:
-    spec: EntrySpec
-    scene: LoadedStepScene | None
-    selector_bundle: SelectorBundle | None = None
-
-
-@dataclass(frozen=True)
-class _CliTargetSpec:
-    target: str
-    output_path: Path | None = None
-
-
-def _cli_progress_line(
-    spec: EntrySpec,
-    *,
-    logger: CliLogger,
-    fallback: str,
-) -> "contextlib.AbstractContextManager[Callable[[ProgressEvent], None] | None]":
-    """:func:`cli_progress_line` keyed to a spec's source ref."""
-    return cli_progress_line(spec.source_ref, logger=logger, fallback=fallback)
-
-
-def _display_name_for_path(path: Path) -> str:
-    return path.stem
-
-
-def _display_path(path: Path) -> str:
-    resolved = path.resolve()
-    try:
-        return resolved.relative_to(Path.cwd().resolve()).as_posix()
-    except ValueError:
-        return resolved.as_posix()
-
-
-def _resolve_cli_output_path(
-    raw_output: str | Path | None,
-    *,
-    expected_suffixes: tuple[str, ...],
-    tool_name: str,
-    option_label: str = "--output",
-) -> Path | None:
-    if raw_output is None:
-        return None
-    value = str(raw_output).strip()
-    if not value:
-        raise ValueError(f"{tool_name} {option_label} must be a non-empty path")
-    if "\\" in value:
-        raise ValueError(f"{tool_name} {option_label} must use POSIX '/' separators")
-    output_path = Path(value).expanduser()
-    resolved = output_path.resolve() if output_path.is_absolute() else (Path.cwd() / output_path).resolve()
-    if resolved.suffix.lower() not in expected_suffixes:
-        joined = " or ".join(expected_suffixes)
-        raise ValueError(f"{tool_name} {option_label} must end in {joined}")
-    return resolved
-
-
-def targets_include_output_pairs(targets: Sequence[str]) -> bool:
-    return any("=" in str(target or "") for target in targets)
-
-
-def _parse_cli_target_specs(
-    targets: Sequence[str],
-    *,
-    expected_suffixes: tuple[str, ...],
-    tool_name: str,
-) -> list[_CliTargetSpec]:
-    specs: list[_CliTargetSpec] = []
-    for target in targets:
-        target_text = str(target or "").strip()
-        if "=" not in target_text:
-            specs.append(_CliTargetSpec(target=target_text))
-            continue
-        raw_source, raw_output = target_text.split("=", 1)
-        source = raw_source.strip()
-        if not source:
-            raise ValueError(f"{tool_name} output pair must use SOURCE=OUTPUT")
-        output_path = _resolve_cli_output_path(
-            raw_output,
-            expected_suffixes=expected_suffixes,
-            tool_name=tool_name,
-            option_label="output pair",
-        )
-        if output_path is None:
-            raise ValueError(f"{tool_name} output pair must use SOURCE=OUTPUT")
-        specs.append(_CliTargetSpec(target=source, output_path=output_path))
-    return specs
-
-
-def _apply_step_options_to_spec(spec: EntrySpec, step_options: StepImportOptions) -> EntrySpec:
-    if not step_options.has_metadata or spec.step_path is None:
-        return spec
-    return replace(
-        spec,
-        mesh_tolerance=step_options.mesh_tolerance if step_options.mesh_tolerance is not None else spec.mesh_tolerance,
-        mesh_angular_tolerance=(
-            step_options.mesh_angular_tolerance
-            if step_options.mesh_angular_tolerance is not None
-            else spec.mesh_angular_tolerance
-        ),
-        mesh_tolerance_explicit=spec.mesh_tolerance_explicit or step_options.mesh_tolerance is not None,
-        mesh_angular_tolerance_explicit=(
-            spec.mesh_angular_tolerance_explicit or step_options.mesh_angular_tolerance is not None
-        ),
-    )
-
-
-def _spec_requests_extra_outputs(spec: EntrySpec) -> bool:
-    """True when the target asks for an on-demand output beyond the render package
-    (`scripts/gen --write-step`). An explicitly requested output must be produced
-    even when the compose is current, so it defeats every no-op and reuse fast
-    path."""
-    return spec.step_export_path is not None
-
-
-def _spec_output_paths(spec: EntrySpec) -> tuple[Path, ...]:
-    paths: list[Path] = []
-    if spec.step_path is not None:
-        paths.append(spec.step_path)
-        paths.append(render_package_dir(spec.entry_path))
-    for path in (spec.dxf_path,):
-        if path is not None:
-            paths.append(path)
-    return tuple(path.resolve() for path in paths)
-
-
-def _validate_cli_output_override(
-    spec: EntrySpec,
-    *,
-    output_path: Path,
-    all_specs: Sequence[EntrySpec],
-    tool_name: str,
-) -> None:
-    resolved_output = output_path.resolve()
-    for candidate in all_specs:
-        if candidate.source_ref == spec.source_ref:
-            continue
-        if resolved_output in _spec_output_paths(candidate):
-            raise ValueError(
-                f"{tool_name} --output would overwrite another CAD output: "
-                f"{_display_path(output_path)} belongs to {candidate.source_ref}"
-            )
-
-
-def _validate_duplicate_cli_output_overrides(
-    output_paths: Sequence[Path | None],
-    *,
-    tool_name: str,
-) -> None:
-    seen: dict[Path, Path] = {}
-    for output_path in output_paths:
-        if output_path is None:
-            continue
-        resolved = output_path.resolve()
-        previous = seen.get(resolved)
-        if previous is not None:
-            raise ValueError(f"{tool_name} output path is used more than once: {_display_path(output_path)}")
-        seen[resolved] = output_path
-
-
-def _apply_step_output_overrides(
-    selected_specs: Sequence[EntrySpec],
-    *,
-    output_paths: Sequence[Path | None],
-    all_specs: Sequence[EntrySpec],
-    tool_name: str,
-) -> list[EntrySpec]:
-    if not any(output_path is not None for output_path in output_paths):
-        return list(selected_specs)
-    if len(output_paths) != len(selected_specs):
-        raise ValueError(f"{tool_name} output override count must match target count")
-    _validate_duplicate_cli_output_overrides(output_paths, tool_name=tool_name)
-    updated_specs: list[EntrySpec] = []
-    for spec, output_path in zip(selected_specs, output_paths, strict=True):
-        if output_path is None:
-            updated_specs.append(spec)
-            continue
-        if spec.source != "generated":
-            raise ValueError(f"{tool_name} output pairs can only be used with generated Python targets")
-        _validate_cli_output_override(spec, output_path=output_path, all_specs=all_specs, tool_name=tool_name)
-        updated_specs.append(
-            replace(
-                spec,
-                cad_ref=cad_ref_from_step_path(output_path),
-                display_name=_display_name_for_path(output_path),
-                step_path=output_path,
-                # A STEP output path is now a STEP *export* request (gen_step writes no STEP
-                # by default): write it on demand to the requested path.
-                step_export_path=output_path,
-            )
-        )
-    return updated_specs
-
-
-def _apply_dxf_output_overrides(
-    selected_specs: Sequence[EntrySpec],
-    *,
-    output_paths: Sequence[Path | None],
-    all_specs: Sequence[EntrySpec],
-    tool_name: str,
-) -> list[EntrySpec]:
-    if not any(output_path is not None for output_path in output_paths):
-        return list(selected_specs)
-    if len(output_paths) != len(selected_specs):
-        raise ValueError(f"{tool_name} output override count must match target count")
-    _validate_duplicate_cli_output_overrides(output_paths, tool_name=tool_name)
-    updated_specs: list[EntrySpec] = []
-    for spec, output_path in zip(selected_specs, output_paths, strict=True):
-        if output_path is None:
-            updated_specs.append(spec)
-            continue
-        if spec.source != "generated":
-            raise ValueError(f"{tool_name} output pairs can only be used with generated Python targets")
-        _validate_cli_output_override(spec, output_path=output_path, all_specs=all_specs, tool_name=tool_name)
-        updated_specs.append(
-            replace(
-                spec,
-                cad_ref=cad_ref_from_dxf_path(output_path),
-                display_name=_display_name_for_path(output_path),
-                # A DXF output path is a DXF *export* request (gen_dxf builds the drawing
-                # package by default): write it on demand to the requested path.
-                dxf_path=output_path,
-                dxf_export_path=output_path,
-            )
-        )
-    return updated_specs
-
-
-def _apply_dxf_output_override(
-    selected_specs: Sequence[EntrySpec],
-    *,
-    output_path: Path | None,
-    all_specs: Sequence[EntrySpec],
-    tool_name: str,
-) -> list[EntrySpec]:
-    if output_path is None:
-        return list(selected_specs)
-    if len(selected_specs) != 1:
-        raise ValueError(f"{tool_name} --output can only be used with exactly one target")
-    spec = selected_specs[0]
-    if spec.source != "generated":
-        raise ValueError(f"{tool_name} --output can only be used with generated Python targets")
-    return _apply_dxf_output_overrides(
-        selected_specs,
-        output_paths=[output_path],
-        all_specs=all_specs,
-        tool_name=tool_name,
-    )
-
-
-def _resolve_discovery_root(root: Path | str) -> Path:
-    candidate = Path(root)
-    resolved = candidate.resolve() if candidate.is_absolute() else (Path.cwd() / candidate).resolve()
-    if not resolved.exists():
-        raise FileNotFoundError(f"CAD discovery directory does not exist: {relative_to_cwd(resolved)}")
-    if not resolved.is_dir():
-        raise NotADirectoryError(f"CAD discovery path is not a directory: {relative_to_cwd(resolved)}")
-    return resolved
-
-
-def list_entry_specs(root: Path | None = None) -> list[EntrySpec]:
-    root = Path.cwd().resolve() if root is None else root
-    specs = [_entry_spec_from_source(source) for source in iter_cad_sources(_resolve_discovery_root(root))]
-    return sorted(specs, key=lambda spec: spec.source_ref)
-
-
-def _entry_spec_from_source(source: CadSource) -> EntrySpec:
-    generator_metadata = source.generator_metadata
-    script_path = source.script_path
-    kind = source.kind
-    step_path = source.step_path
-    mesh_settings = resolve_mesh_settings(
-        cad_ref=source.cad_ref,
-        generator_metadata=generator_metadata,
-        mesh_tolerance=source.mesh_tolerance,
-        mesh_angular_tolerance=source.mesh_angular_tolerance,
-    )
-    display_path = step_path if step_path is not None else source.source_path
-
-    return EntrySpec(
-        source_ref=source.source_ref,
-        cad_ref=source.cad_ref,
-        kind=kind,
-        source_path=source.source_path,
-        display_name=(
-            generator_metadata.display_name
-            if generator_metadata is not None and generator_metadata.display_name
-            else _display_name_for_path(display_path)
-        ),
-        source=source.source,
-        step_path=step_path,
-        script_path=script_path,
-        generator_metadata=generator_metadata,
-        dxf_path=source.dxf_path,
-        mesh_tolerance=mesh_settings.tolerance,
-        mesh_angular_tolerance=mesh_settings.angular_tolerance,
-        mesh_tolerance_explicit=source.mesh_tolerance is not None,
-        mesh_angular_tolerance_explicit=source.mesh_angular_tolerance is not None,
-        color=source.color,
-    )
-
-
-def selected_entry_specs(all_specs: Sequence[EntrySpec], source_refs: Sequence[str]) -> list[EntrySpec]:
-    if not source_refs:
-        raise ValueError("At least one CAD target is required")
-    by_source = {spec.source_ref: spec for spec in all_specs}
-    by_cad_ref = {spec.cad_ref: spec for spec in all_specs}
-    by_step_path = {
-        spec.step_path.resolve(): spec
-        for spec in all_specs
-        if spec.step_path is not None
-    }
-    selected: list[EntrySpec] = []
-    for source_ref in source_refs:
-        spec = _spec_for_source_ref(source_ref, by_source=by_source, by_cad_ref=by_cad_ref, by_step_path=by_step_path)
-        if spec is None:
-            raise FileNotFoundError(f"CAD source not found: {source_ref}")
-        selected.append(spec)
-    return selected
-
-
-def _spec_for_source_ref(
-    raw_ref: str,
-    *,
-    by_source: dict[str, EntrySpec],
-    by_cad_ref: dict[str, EntrySpec],
-    by_step_path: dict[Path, EntrySpec],
-) -> EntrySpec | None:
-    source_ref = normalize_source_ref(raw_ref)
-    if source_ref and source_ref in by_source:
-        return by_source[source_ref]
-    cad_ref = normalize_cad_ref(raw_ref)
-    if cad_ref and cad_ref in by_cad_ref:
-        return by_cad_ref[cad_ref]
-    candidate = Path(str(raw_ref or "").strip())
-    if candidate:
-        resolved = candidate.resolve() if candidate.is_absolute() else (
-            Path.cwd() / candidate
-        )
-        resolved = resolved.resolve()
-        if resolved in by_step_path:
-            return by_step_path[resolved]
-        source = find_source_by_path(resolved)
-        if source is not None:
-            return by_source.get(source.source_ref)
-    return None
-
-
-def _mesh_tolerance_is_explicit(spec: EntrySpec) -> bool:
-    return bool(spec.mesh_tolerance_explicit) or not math.isclose(
-        float(spec.mesh_tolerance),
-        float(DEFAULT_MESH_TOLERANCE),
-        rel_tol=1e-12,
-        abs_tol=1e-12,
-    )
-
-
-def _mesh_angular_tolerance_is_explicit(spec: EntrySpec) -> bool:
-    return bool(spec.mesh_angular_tolerance_explicit) or not math.isclose(
-        float(spec.mesh_angular_tolerance),
-        float(DEFAULT_MESH_ANGULAR_TOLERANCE),
-        rel_tol=1e-12,
-        abs_tol=1e-12,
-    )
-
-
-# How much finer than the adaptive floor an explicit tolerance has to be before
-# it is almost certainly a mistake rather than a deliberate choice.
-_TOLERANCE_WARN_RATIO = 8.0
-
-
-def _warn_if_tolerance_defeats_scale_floor(spec: EntrySpec, adaptive: object) -> None:
-    """Say something when ``--mesh-tolerance`` silently defeats the size floor.
-
-    For anything larger than desk scale the adaptive resolver floors the linear
-    deflection proportionally to the model diagonal, because meshing a metre-scale
-    part at micron-class chord error costs minutes and hundreds of megabytes. An
-    explicit ``--mesh-tolerance`` overrides that floor completely — which is
-    correct, but silent, and 0.02 looks like a safe "default" value to pass. On a
-    5.4 m car it is 80x finer than the floor and turns a 15-second build into a
-    six-minute one with nothing in the output to explain why.
-    """
-    settings = getattr(adaptive, "settings", None)
-    floor = float(getattr(settings, "tolerance", 0.0) or 0.0)
-    requested = float(spec.mesh_tolerance)
-    if floor <= 0.0 or requested <= 0.0 or requested >= floor / _TOLERANCE_WARN_RATIO:
-        return
-    hints = getattr(adaptive, "hints", None)
-    diagonal = 0.0
-    if isinstance(hints, Mapping):
-        raw = hints.get("bboxDiag")
-        diagonal = float(raw) if isinstance(raw, (int, float)) else 0.0
-    print(
-        f"[cadgen] warning: --mesh-tolerance {requested:g} mm is "
-        f"{floor / requested:.0f}x finer than the {floor:g} mm this model's size "
-        f"({diagonal:.0f} mm diagonal) would otherwise use. Meshing will be much "
-        f"slower and the package much larger. Omit --mesh-tolerance to let it scale.",
-        file=sys.stderr,
-        flush=True,
-    )
-
-
-def _selector_options_for_part(spec: EntrySpec, *, scene: LoadedStepScene | None = None) -> SelectorOptions:
-    defaults = SelectorOptions()
-    linear_deflection = spec.mesh_tolerance
-    angular_deflection = spec.mesh_angular_tolerance
-    resolution: dict[str, object] = {
-        "mode": "explicit",
-        "profile": "custom",
-        "linearExplicit": True,
-        "angularExplicit": True,
-    }
-    linear_explicit = _mesh_tolerance_is_explicit(spec)
-    angular_explicit = _mesh_angular_tolerance_is_explicit(spec)
-    edge_visibility_classes = normalize_step_edge_render_visibility_classes(None)
-    if isinstance(scene, LoadedStepScene):
-        adaptive = adaptive_mesh_resolution_for_scene(scene)
-        if not linear_explicit:
-            linear_deflection = adaptive.settings.tolerance
-        else:
-            _warn_if_tolerance_defeats_scale_floor(spec, adaptive)
-        if not angular_explicit:
-            angular_deflection = adaptive.settings.angular_tolerance
-        edge_visibility_classes = _edge_visibility_classes_for_resolution(adaptive.profile, adaptive.hints)
-        resolution = {
-            "mode": "auto",
-            "profile": adaptive.profile,
-            "linearExplicit": linear_explicit,
-            "angularExplicit": angular_explicit,
-            "hints": adaptive.hints,
-        }
-    return SelectorOptions(
-        linear_deflection=linear_deflection,
-        angular_deflection=angular_deflection,
-        relative=defaults.relative,
-        edge_deflection=defaults.edge_deflection,
-        edge_deflection_ratio=defaults.edge_deflection_ratio,
-        max_edge_points=defaults.max_edge_points,
-        digits=defaults.digits,
-        mesh_resolution=resolution,
-        edge_visibility_classes=edge_visibility_classes,
-    )
-
-
-def _edge_visibility_classes_for_resolution(profile: str, hints: Mapping[str, object] | None) -> tuple[str, ...]:
-    normalized_profile = str(profile or "").strip().lower()
-    hint_values = hints if isinstance(hints, Mapping) else {}
-    occurrence_edge_count = _hint_int(hint_values.get("occurrenceEdgeCount"))
-    feature_only = (
-        normalized_profile in {"large-topology", "coarse-assembly"}
-        or occurrence_edge_count >= 8000
-    )
-    if feature_only:
-        return (STEP_EDGE_VISIBILITY_CLASSES["FEATURE"],)
-    return normalize_step_edge_render_visibility_classes(None)
-
-
-def _hint_float(value: object) -> float:
-    try:
-        return float(value or 0.0)
-    except (TypeError, ValueError):
-        return 0.0
-
-
-def _hint_int(value: object) -> int:
-    return int(_hint_float(value))
-
-
-def _load_generator_module(script_path: Path) -> object:
-    resolved_script_path = script_path.resolve()
-    module_name = (
-        "_cad_tool_"
-        + _display_path(resolved_script_path).replace("/", "_").replace("\\", "_").replace("-", "_").replace(".", "_")
-    )
-    module_spec = importlib.util.spec_from_file_location(module_name, resolved_script_path)
-    if module_spec is None or module_spec.loader is None:
-        raise RuntimeError(f"Failed to load generator module from {_display_path(resolved_script_path)}")
-
-    module = importlib.util.module_from_spec(module_spec)
-    original_sys_path = list(sys.path)
-    # Seed sys.path so the generator's module-top imports (its sibling/shared packages such as
-    # robot_common / STEP) resolve. Derive everything from the generator script's OWN location —
-    # its folder, plus any ancestor that is a package root (contains a STEP/ or robot_common/
-    # package) — so resolution is independent of the process working directory. Deliberately NOT
-    # seeding the repo root or skills/cad/scripts: a generator must not depend on the repository's
-    # skills/ being importable (AGENTS.md skill isolation).
-    search_paths = [str(resolved_script_path.parent)]
-    for parent in resolved_script_path.parents:
-        if (
-            (parent / "STEP" / "__init__.py").is_file()
-            or (parent / "robot_common" / "__init__.py").is_file()
-        ):
-            search_paths.append(str(parent))
-    for candidate in reversed(search_paths):
-        if candidate not in sys.path:
-            sys.path.insert(0, candidate)
-
-    try:
-        sys.modules[module_name] = module
-        module_spec.loader.exec_module(module)
-    finally:
-        sys.path[:] = original_sys_path
-
-    return module
-
-
-def _resolve_params_sidecar(params: object, *, script_path: Path) -> Path:
-    """Resolve the optional gen_step() ``params`` value — a filepath to a hand-authored
-    JS sidecar (the step-module manifest: parameters/features/animations/update) — to an
-    absolute path. A relative path is resolved against the generator file's directory.
-    The sidecar is hand-authored source, so it must already exist on disk."""
-    if not isinstance(params, (str, Path)):
-        raise TypeError(
-            f"{_display_path(script_path)} gen_step() envelope field 'params' must be a "
-            f"filepath string, got {type(params).__name__}"
-        )
-    candidate = Path(params)
-    resolved = (candidate if candidate.is_absolute() else script_path.parent / candidate).resolve()
-    if not resolved.is_file():
-        raise FileNotFoundError(
-            f"{_display_path(script_path)} gen_step() params sidecar not found: {params}"
-        )
-    return resolved
-
-
-def _normalize_step_payload(
-    result: object,
-    *,
-    script_path: Path,
-) -> dict[str, object]:
-    from build123d import Shape as Build123dShape
-
-    if isinstance(result, Build123dShape):
-        return {"shape": result}
-    if isinstance(result, dict):
-        # stl / 3mf / mesh_tolerance / mesh_angular_tolerance are consumed via the static
-        # metadata path (per-generator STL/3MF outputs + mesh tolerances); keep allowing them.
-        allowed_fields = {"shape", "params", "stl", "3mf", "mesh_tolerance", "mesh_angular_tolerance"}
-        extra_fields = sorted(str(key) for key in result if key not in allowed_fields)
-        if extra_fields:
-            joined = ", ".join(extra_fields)
-            raise TypeError(f"{_display_path(script_path)} gen_step() envelope has unsupported field(s): {joined}")
-        if "shape" not in result:
-            raise TypeError(
-                f"{_display_path(script_path)} gen_step() envelope must define 'shape'"
-            )
-        envelope: dict[str, object] = {"shape": result["shape"]}
-        params = result.get("params")
-        if params is not None:
-            envelope["params"] = _resolve_params_sidecar(params, script_path=script_path)
-        return envelope
-    raise TypeError(
-        f"{_display_path(script_path)} gen_step() must return a build123d Shape "
-        "or a {'shape': ..., 'params': ...} envelope"
-    )
-
-
-def _normalize_dxf_payload(result: object, *, script_path: Path) -> dict[str, object]:
-    if isinstance(result, dict):
-        allowed_fields = {"document"}
-        extra_fields = sorted(str(key) for key in result if key not in allowed_fields)
-        if extra_fields:
-            joined = ", ".join(extra_fields)
-            raise TypeError(f"{_display_path(script_path)} gen_dxf() envelope has unsupported field(s): {joined}")
-        if "document" not in result:
-            raise TypeError(f"{_display_path(script_path)} gen_dxf() envelope must define 'document'")
-        return {"document": result["document"]}
-    return {"document": result}
-
-
-def _shape_payload_entry_kind(shape: object, *, fallback: str) -> str:
-    if fallback not in {"part", "assembly"}:
-        raise RuntimeError(f"Unsupported generated STEP kind: {fallback}")
-    if (
-        fallback == "assembly"
-        or _shape_has_explicit_children(shape)
-        or _shape_is_multi_child_compound(shape)
-    ):
-        return "assembly"
-    return "part"
-
-
-def _shape_has_explicit_children(shape: object) -> bool:
-    try:
-        from build123d import Shape as Build123dShape
-    except Exception:
-        return False
-    if not isinstance(shape, Build123dShape):
-        return False
-    try:
-        return bool(tuple(getattr(shape, "children", ()) or ()))
-    except TypeError:
-        return False
-
-
-def _shape_is_multi_child_compound(shape: object) -> bool:
-    try:
-        from OCP.TopAbs import TopAbs_COMPOUND
-        from OCP.TopoDS import TopoDS_Iterator
-        from build123d import Shape as Build123dShape
-    except Exception:
-        return False
-    if not isinstance(shape, Build123dShape):
-        return False
-    wrapped = getattr(shape, "wrapped", None)
-    if wrapped is None:
-        return False
-    try:
-        if wrapped.ShapeType() != TopAbs_COMPOUND:
-            return False
-    except Exception:
-        return False
-    iterator = TopoDS_Iterator(wrapped)
-    count = 0
-    while iterator.More():
-        count += 1
-        if count > 1:
-            return True
-        iterator.Next()
-    return False
-
-
-def _mark_scene_step_payload(
-    scene: LoadedStepScene,
-    *,
-    entry_kind: str,
-    payload_kind: str,
-) -> LoadedStepScene:
-    if isinstance(scene, LoadedStepScene):
-        scene.text_to_cad_entry_kind = entry_kind
-        scene.step_payload_kind = payload_kind
-    return scene
-
-
-def _scene_entry_kind(scene: LoadedStepScene | None) -> str | None:
-    if scene is None:
-        return None
-    entry_kind = str(getattr(scene, "text_to_cad_entry_kind", "") or "").strip().lower()
-    return entry_kind if entry_kind in {"part", "assembly"} else None
-
-
-def _effective_step_spec_for_scene(spec: EntrySpec, scene: LoadedStepScene | None) -> EntrySpec:
-    entry_kind = _scene_entry_kind(scene)
-    if entry_kind is None or entry_kind == spec.kind:
-        return spec
-    return replace(spec, kind=entry_kind)
-
-
-def _write_shape_step_payload(
-    envelope: dict[str, object],
-    *,
-    output_path: Path,
-    script_path: Path,
-    logger: CliLogger,
-    entry_kind: str,
-) -> LoadedStepScene:
-    shape = envelope.get("shape")
-    from build123d import Shape as Build123dShape
-
-    if not isinstance(shape, Build123dShape):
-        raise TypeError(
-            f"{_display_path(script_path)} gen_step() envelope field 'shape' must be a build123d Shape, "
-            f"got {type(shape).__name__}"
-        )
-    # gen_step builds the render scene in memory and does NOT write a text STEP — STEP is
-    # written on demand from scene.source_compound (`scripts/gen --write-step`, or the
-    # Viewer's Save-dialog export). The scene is built straight from the XCAF doc, never
-    # via a STEP round-trip.
-    source_identity = python_source_hash(script_path)
-    scene = build_build123d_step_scene(
-        shape,
-        output_path,
-        source_kind="python",
-        source_hash=source_identity.source_hash,
-    )
-    _mark_scene_python_backed(scene, source_identity=source_identity, source_path=script_path)
-    _mark_scene_step_payload(scene, entry_kind=entry_kind, payload_kind="shape")
-    # Stash the pre-bake compound: the component-package emit job introspects its located
-    # children (occurrence transforms + dedup), and the STEP export serializes it.
-    scene.source_compound = shape
-    params_abs = envelope.get("params")
-    if params_abs is not None:
-        # Record the hand-authored JS sidecar model-folder-relative, like sourcePath, so the
-        # descriptor stays portable. The viewer reads this back from assembly.json.
-        scene.params_path = relative_to_file(Path(params_abs), scene.step_path)
-    logger.debug(f"built render scene (no STEP written): {_display_path(output_path)}")
-    return scene
-
-
-def _mark_scene_python_backed(
-    scene: LoadedStepScene,
-    *,
-    source_identity: PythonSourceHash,
-    source_path: Path,
-) -> LoadedStepScene:
-    if not isinstance(scene, LoadedStepScene):
-        return scene
-    scene.source_kind = "python"
-    scene.source_hash = source_identity.source_hash
-    scene.source_path = relative_to_file(source_path, scene.step_path)
-    return scene
-
-
-def _write_dxf_payload(
-    envelope: dict[str, object],
-    *,
-    output_path: Path,
-    script_path: Path,
-    logger: CliLogger,
-) -> None:
-    document = envelope.get("document")
-    saveas = getattr(document, "saveas", None)
-    if not callable(saveas):
-        raise TypeError(
-            f"{_display_path(script_path)} gen_dxf() envelope field 'document' must be a DXF document, "
-            f"got {type(document).__name__}"
-        )
-    output_path.parent.mkdir(parents=True, exist_ok=True)
-    saveas(str(output_path))
-    source_identity = python_source_hash(script_path)
-    write_dxf_text_to_cad_metadata(
-        output_path,
-        text_to_cad_identity_metadata(
-            source_path=relative_to_file(script_path, output_path),
-            source_hash=source_identity.source_hash,
-        ),
-    )
-    logger.debug(f"wrote DXF: {_display_path(output_path)}")
-
-
-def run_script_generator(
-    spec: EntrySpec,
-    generator_name: str,
-    *,
-    logger: CliLogger | None = None,
-    force: bool = False,
-    reset_runtime_closure: bool = False,
-    progress: object | None = None,
-    lock_intent: str = "write",
-) -> LoadedStepScene | None:
-    """Run a generator's ``gen_step``/``gen_dxf`` and return its scene.
-
-    ``lock_intent`` says whether this run will rewrite the model's render package
-    (``"write"``, the default) or merely occupy its generator (``"generate"`` -- an export,
-    a topology extraction, an interference check). See :func:`_track_spec_generation`:
-    getting this wrong makes an export look like a build to the CAD Viewer.
-
-    Closure capture is deterministic in every process shape: first-party modules
-    are evicted from ``sys.modules`` BEFORE the generator loads (so its full
-    dependency closure is freshly imported on every run — warm worker, multi-target
-    CLI loop, or cold process alike, and regardless of earlier failed builds), and
-    every first-party file EXECUTED during the run is recorded via the ``exec``
-    audit event (so dependencies survive even when a generator unloads modules
-    from ``sys.modules`` mid-run). Only first-party ``.py`` modules are evicted
-    (see :func:`repo_local_loaded_modules`); the running runtime (cadgen, the CLI
-    launcher) and C extensions / site-packages (numpy, OCP, build123d) are never
-    touched — they cannot reload, must stay warm, and are not freshness inputs.
-
-    ``reset_runtime_closure`` is retained for API compatibility (the warm worker
-    passes it) but is now a no-op: the pre-run eviction supersedes it.
-    """
-    logger = logger or CliLogger("cad")
-    if generator_name not in {"gen_step", "gen_dxf"}:
-        raise RuntimeError(f"Unsupported generator: {generator_name}")
-    if spec.script_path is None or spec.generator_metadata is None:
-        raise ValueError(f"{spec.source_ref} is not a generated Python CAD source")
-    # A WRITER arrives with the BuildRun that already owns this model's status record and
-    # its progress line. An EXPORT arrives with neither: it takes the generator lock instead
-    # of the write lock, and until that lock carried a reporter, `cad export` ran the same
-    # multi-minute gen_step() a build runs and said nothing on any surface. So the run the
-    # lock yields becomes the reporter when nobody above us is one.
-    owns_reporting = progress is None
-    with _generator_progress_line(spec, logger=logger, active=owns_reporting) as sink:
-        with _track_spec_generation(
-            spec, generator_name, intent=lock_intent, logger=logger, sink=sink
-        ) as generator_run:
-            active = generator_run if owns_reporting else progress
-            # The phase opens INSIDE the lock: before this it opened first, so a run queued
-            # behind a peer reported "building geometry" for the whole time it was waiting.
-            resolve_progress(active).phase(PHASE_GENERATE)
-            return _run_script_generator_inner(
-                spec,
-                generator_name,
-                logger=logger,
-                force=force,
-                reset_runtime_closure=reset_runtime_closure,
-                progress=active,
-            )
-
-
-@contextlib.contextmanager
-def _generator_progress_line(
-    spec: EntrySpec, *, logger: CliLogger | None, active: bool
-) -> Iterator[Callable[[ProgressEvent], None] | None]:
-    """The terminal line for a generator run that owns its own reporting.
-
-    Inactive when a build above us already paints one — two painters on one tty interleave
-    into nonsense — and when there is no logger to paint through."""
-    if not active:
-        yield None
-        return
-    with cli_progress_line(
-        spec.source_ref, logger=logger or CliLogger("cad"), fallback="Building..."
-    ) as sink:
-        yield sink
-
-
-def _run_script_generator_inner(
-    spec: EntrySpec,
-    generator_name: str,
-    *,
-    logger: CliLogger,
-    force: bool = False,
-    reset_runtime_closure: bool = False,
-    progress: object | None = None,
-) -> LoadedStepScene | None:
-    del reset_runtime_closure  # superseded by the pre-run eviction below; kept for API compat
-    generated_scene: LoadedStepScene | None = None
-    # Deterministic closure capture (see run_script_generator's docstring): start from a
-    # clean first-party module space, then record every first-party file executed while
-    # the generator loads and runs. The recorded set is complete even if the generator
-    # unloads modules mid-run; the sys.modules delta stays as a belt-and-braces union.
-    evict_first_party_modules()
-    modules_before_load = set(sys.modules)
-    with record_first_party_execution() as executed_files:
-        with logger.timed(f"load generator {spec.source_ref}"):
-            module = _load_generator_module(spec.script_path)
-        generator = getattr(module, generator_name, None)
-        if not callable(generator):
-            raise RuntimeError(f"{_display_path(spec.script_path)} does not define callable {generator_name}()")
-        # Bind the lock holder as the ambient reporter for the generator's own code. This is
-        # the in-process twin of `run_node_builder`, which lets a Node child describe its
-        # work over a pipe: gen_step() takes no arguments and so cannot be handed the run,
-        # and without this the longest phase of most builds reports nothing at all. Silent
-        # generators are unaffected -- nothing reads the binding unless they ask for it.
-        with logger.timed(f"run {generator_name} {spec.source_ref}"), reporting_as(progress):
-            raw_payload = generator()
-
-    source_closure: PythonSourceClosure | None = None
-    if generator_name == "gen_step":
-        envelope = _normalize_step_payload(raw_payload, script_path=spec.script_path)
-        if spec.step_path is None:
-            raise RuntimeError(f"{spec.source_ref} has no configured STEP output")
-        # Record paths relative to the model folder so the descriptor stays portable.
-        source_closure = capture_runtime_closure(
-            modules_before_load,
-            spec.script_path,
-            base=spec.step_path.parent,
-            executed_files=executed_files,
-        )
-        generated_scene = _write_shape_step_payload(
-            envelope,
-            output_path=spec.step_path,
-            script_path=spec.script_path,
-            logger=logger,
-            entry_kind=_shape_payload_entry_kind(envelope.get("shape"), fallback=spec.kind),
-        )
-    elif generator_name == "gen_dxf":
-        from cadgen._internal.drawing_package import write_drawing_package
-        from cadgen.drawing_checks import raise_on_error_findings, validate_drawing_document
-
-        envelope = _normalize_dxf_payload(raw_payload, script_path=spec.script_path)
-        if spec.dxf_path is None:
-            raise RuntimeError(f"{spec.source_ref} has no configured DXF output")
-        # Validation happens IN generation: the in-memory document is checked once,
-        # gating the drawing package and every export alike. Fail closed — anything
-        # that is not a real drawing document is rejected rather than skipped.
-        document = envelope.get("document")
-        if not (hasattr(document, "modelspace") and hasattr(document, "header")):
-            raise TypeError(
-                f"{_display_path(spec.script_path)} gen_dxf() must return an ezdxf "
-                f"document, got {type(document).__name__}"
-            )
-        findings = validate_drawing_document(document)
-        for finding in findings:
-            if finding.severity != "error":
-                logger.info(f"{spec.source_ref} {finding.render()}")
-        raise_on_error_findings(findings, label=_display_path(spec.script_path))
-        # Mirror gen_step: capture the generator's closure (relative to the model
-        # folder) so the drawing package records the freshness inputs the viewer's
-        # staleness gate reads. Code reuse is the freshness link: a drawing that
-        # path-loads its .step.py records it (and its imports) here. Non-Python inputs
-        # (e.g. an imported vendor .step) are intentionally NOT tracked — like a
-        # gen_step that composes imported STEPs, the drawing does not rebuild when
-        # they change. Then the default build product is the drawing package; the
-        # sibling/exported .dxf is written on demand only.
-        source_closure = capture_runtime_closure(
-            modules_before_load,
-            spec.script_path,
-            base=spec.script_path.parent,
-            executed_files=executed_files,
-        )
-        # `progress` is the BuildRun holding this package's lock. It is threaded down here
-        # because writing the package now includes baking preview.glb in a Node child, and
-        # that child reports its phases -- and proves its run id against the lock sentinel --
-        # through the very run that is holding the lock (design §4.3, §7.4.2).
-        write_drawing_package(
-            envelope.get("document"),
-            script_path=spec.script_path,
-            source_closure=source_closure,
-            run=progress,
-        )
-        logger.debug(
-            f"wrote drawing package: {_display_path(render_package_dir(spec.script_path))}"
-        )
-        if spec.dxf_export_path is not None:
-            _write_dxf_payload(
-                envelope, output_path=spec.dxf_export_path, script_path=spec.script_path, logger=logger
-            )
-    if generated_scene is not None and source_closure is not None:
-        generated_scene.source_closure_hash = source_closure.closure_hash
-        generated_scene.source_closure_files = source_closure.files
-    if (
-        generator_name == "gen_dxf"
-        and spec.dxf_export_path is not None
-        and not spec.dxf_export_path.exists()
-    ):
-        raise RuntimeError(
-            f"{_display_path(spec.script_path)} did not write {_display_path(spec.dxf_export_path)}"
-        )
-    return generated_scene if generator_name == "gen_step" else None
-
-
-def _is_git_lfs_pointer(step_path: Path) -> bool:
-    try:
-        with step_path.open("rb") as handle:
-            return handle.read(len(GIT_LFS_POINTER_PREFIX)) == GIT_LFS_POINTER_PREFIX
-    except OSError:
-        return False
-
-
-def _ensure_step_ready(step_path: Path) -> None:
-    if not step_path.exists():
-        raise FileNotFoundError(f"STEP file is missing: {_display_path(step_path)}")
-    if _is_git_lfs_pointer(step_path):
-        raise RuntimeError(
-            f"{_display_path(step_path)} is a Git LFS pointer, not the real STEP file.\n"
-            "Fetch Git LFS objects before generating CAD artifacts.\n"
-            "For Vercel Git deployments, enable Git LFS in Project Settings > Git and redeploy."
-        )
-
-
-@dataclass(frozen=True)
-class _ArtifactJob:
-    name: str
-    run: Callable[[], object]
-
-
-def _run_artifact_jobs(
-    jobs: Sequence[_ArtifactJob],
-    *,
-    logger: CliLogger | None = None,
-) -> dict[str, object]:
-    results: dict[str, object] = {}
-    for job in jobs:
-        if logger is not None:
-            with logger.timed(f"write {job.name}"):
-                results[job.name] = job.run()
-        else:
-            results[job.name] = job.run()
-    return results
-
+from cadgen._internal.generation_runner import (
+    GIT_LFS_POINTER_PREFIX,
+    _ArtifactJob,
+    _effective_step_spec_for_scene,
+    _ensure_step_ready,
+    _generator_progress_line,
+    _is_git_lfs_pointer,
+    _load_generator_module,
+    _mark_scene_python_backed,
+    _mark_scene_step_payload,
+    _normalize_dxf_payload,
+    _normalize_step_payload,
+    _resolve_params_sidecar,
+    _run_artifact_jobs,
+    _run_script_generator_inner,
+    _scene_entry_kind,
+    _shape_has_explicit_children,
+    _shape_is_multi_child_compound,
+    _shape_payload_entry_kind,
+    _spec_output_dir,
+    _track_spec_generation,
+    _write_dxf_payload,
+    _write_lock_without_reporting,
+    _write_shape_step_payload,
+    run_script_generator,
+)
+from cadgen._internal.generation_spec import (
+    EntrySpec,
+    GeneratedStepResult,
+    _CliTargetSpec,
+    _TOLERANCE_WARN_RATIO,
+    _apply_dxf_output_override,
+    _apply_dxf_output_overrides,
+    _apply_step_options_to_spec,
+    _apply_step_output_overrides,
+    _cli_progress_line,
+    _display_name_for_path,
+    _display_path,
+    _edge_visibility_classes_for_resolution,
+    _entry_spec_from_source,
+    _hint_float,
+    _hint_int,
+    _mesh_angular_tolerance_is_explicit,
+    _mesh_tolerance_is_explicit,
+    _parse_cli_target_specs,
+    _resolve_cli_output_path,
+    _resolve_discovery_root,
+    _selector_options_for_part,
+    _spec_for_source_ref,
+    _spec_output_paths,
+    _spec_requests_extra_outputs,
+    _validate_cli_output_override,
+    _validate_duplicate_cli_output_overrides,
+    _warn_if_tolerance_defeats_scale_floor,
+    list_entry_specs,
+    selected_entry_specs,
+    targets_include_output_pairs,
+)
 
 def _mesh_values_match(
     mesh: Mapping[str, object],
@@ -1764,61 +828,6 @@ class _ContendedGeneration:
 
     def __init__(self, spec: EntrySpec) -> None:
         self.spec = spec
-
-
-def _spec_output_dir(spec: EntrySpec, generator_name: str) -> Path | None:
-    """The coordinated output directory for this spec's generator, if it has one."""
-    if generator_name == "gen_step" and spec.step_path is not None:
-        return render_package_dir(spec.entry_path)
-    if generator_name == "gen_dxf" and spec.script_path is not None:
-        return render_package_dir(spec.script_path)
-    return None
-
-
-def _track_spec_generation(
-    spec: EntrySpec,
-    generator_name: str,
-    *,
-    intent: str = "write",
-    logger: CliLogger | None = None,
-    sink: Callable[[ProgressEvent], None] | None = None,
-) -> contextlib.AbstractContextManager[object]:
-    """Coordinate a generator run against the model's render package.
-
-    ``intent`` picks the SENTINEL, and the distinction is the whole point of there being
-    two. A run that will rewrite the package takes the writer lock, which makes a reader
-    hide the artifact and show a build. A run that merely OCCUPIES the generator and
-    writes the package nothing -- an export, an on-demand topology extraction, an
-    interference check -- takes the generator lock instead. Taking the writer lock for
-    those made a fully-current model report `generating` with an empty bar for the whole
-    length of an export.
-
-    The two sentinels are different files, so they do NOT exclude each other: a build and
-    an export of one model each run its ``gen_step()``, concurrently, in separate
-    processes. That is duplicated work rather than a hazard (no shared in-process state,
-    different outputs), and it is the price of letting a reader tell "being rewritten"
-    from "generator busy" -- see :func:`cadgen.coordination.generator_busy`.
-    """
-    output_dir = _spec_output_dir(spec, generator_name)
-    if output_dir is None:
-        return contextlib.nullcontext()
-    on_wait = lock_wait_notice(logger, spec.source_ref)
-    if intent == "generate":
-        # The kind decides which phase set the run reports over, so a drawing generator
-        # counts its own phases rather than a STEP package's.
-        kind = DRAWING_PACKAGE if generator_name == "gen_dxf" else STEP_PACKAGE
-        return generator_busy(kind, output_dir, on_wait=on_wait, sink=sink)
-    # A writer already has its BuildRun from artifact_build; this only needs the lock, and
-    # yields None so the caller's `progress or this` choice stays a simple one.
-    return _write_lock_without_reporting(write_lock_path(output_dir), on_wait=on_wait)
-
-
-@contextlib.contextmanager
-def _write_lock_without_reporting(
-    path: Path, *, on_wait: Callable[[float], None] | None
-) -> Iterator[None]:
-    with exclusive(path, on_wait=on_wait):
-        yield None
 
 
 def _run_with_spec_generation_status(
